🐛 fix(manifest.json): update default_popup value The default_popup value in the browser_action object has been updated to "www/index.html" to fix the issue where the popup was not displaying correctly.
🔥 chore(FakerPage.vue): remove unused FakerPage component
The FakerPage component has been removed from the project, and its associated file has been deleted.

🎨 style(IndexPage.vue): adjust layout and styles
💄 chore(IndexPage.vue): remove unused styles and elements
The layout of the IndexPage.vue has been adjusted to improve the display of the list of faker methods. The minimum width of the page has been set to 400px to ensure that the list is displayed properly. The styles for the main background and footer have been removed as they are no longer needed. Unused elements such as the main title and the download button have been removed.
